refactor: deepen CAPTCHA retry coordination
