+++ src/lib/automation/server/runner.check.ts
@@ -2,7 +2,6 @@ import assert from "node:assert/strict";
 import test from "node:test";
 import {
   chmodSync,
-  existsSync,
   mkdirSync,
   mkdtempSync,
   readFileSync,
@@ -127,91 +126,6 @@ async function settleStartedBatchTasks(taskIds: readonly string[]) {
   }
 }
 
-function writeTypedCaptchaRoundFixture(rootDir: string, options: {
-  cleanupStartedPath?: string;
-  allowCleanupPath?: string;
-}) {
-  const roundScriptPath = join(rootDir, "captcha-round.mjs");
-  writeFileSync(
-    roundScriptPath,
-    `import { appendFileSync } from "node:fs";
-import { createConnection } from "node:net";
-
-const endpoint = process.env.OCTOPUSBEAK_CAPTCHA_ROUND_OUTCOME_ENDPOINT;
-const token = process.env.OCTOPUSBEAK_CAPTCHA_ROUND_OUTCOME_TOKEN;
-const executionId = process.env.OCTOPUSBEAK_CAPTCHA_ROUND_OUTCOME_EXECUTION_ID;
-appendFileSync(process.env.CAPTCHA_LAUNCH_PATH, executionId + "\\n");
-const frame = (value) => JSON.stringify(value) + "\\n";
-const auth = frame({
-  schemaVersion: 1,
-  type: "authenticate",
-  executionId,
-  token,
-});
-const captured = frame({
-  schemaVersion: 1,
-  type: "captcha-round-outcome",
-  executionId,
-  sequence: 1,
-  outcome: {
-    kind: "captured",
-    stageId: "fubon-login-captcha",
-    challengeKind: "text-captcha",
-  },
-});
-const exhausted = frame({
-  schemaVersion: 1,
-  type: "captcha-round-outcome",
-  executionId,
-  sequence: 2,
-  outcome: {
-    kind: "retryable",
-    stageId: "fubon-login-captcha",
-    reason: "solver-exhausted",
-  },
-});
-await new Promise((resolve, reject) => {
-  const socket = createConnection(endpoint);
-  socket.once("connect", () => socket.end(auth + captured + exhausted, resolve));
-  socket.once("error", reject);
-});
-`,
-    "utf8",
-  );
-  const npmPath = join(rootDir, "bin", "npm");
-  writeFileSync(
-    npmPath,
-    "#!/bin/sh\nnode \"$CAPTCHA_ROUND_SCRIPT\"\n",
-    "utf8",
-  );
-  chmodSync(npmPath, 0o755);
-  const npxPath = join(rootDir, "bin", "npx");
-  writeFileSync(
-    npxPath,
-    [
-      "#!/bin/sh",
-      "if [ \"$1\" = \"libretto\" ] && [ \"$2\" = \"close\" ]; then",
-      options.cleanupStartedPath ? `  touch \"${options.cleanupStartedPath}\"` : "  :",
-      options.allowCleanupPath
-        ? `  while [ ! -f \"${options.allowCleanupPath}\" ]; do sleep 0.01; done`
-        : "  :",
-      "fi",
-      "exit 0",
-      "",
-    ].join("\n"),
-    "utf8",
-  );
-  chmodSync(npxPath, 0o755);
-  return roundScriptPath;
-}
-
-async function waitForFile(path: string) {
-  for (let attempt = 0; attempt < 200; attempt += 1) {
-    if (existsSync(path)) return;
-    await new Promise((resolve) => setTimeout(resolve, 10));
-  }
-  throw new Error(`Timed out waiting for ${path}`);
-}
 test("manual Fubon starts ignore persisted statement selection", async () => {
   const dir = mkdtempSync(join(tmpdir(), "automation-start-selection-"));
   const originalCwd = process.cwd();
@@ -870,125 +784,6 @@ test("cleanup failure makes a partial Libretto run failed", async () => {
     process.chdir(previousCwd);
     if (previousPath === undefined) delete process.env.PATH;
     else process.env.PATH = previousPath;
-    rmSync(rootDir, { recursive: true, force: true });
-  }
-});
-
-test("typed solver exhaustion restarts fresh sessions serially in one task run", async () => {
-  const rootDir = mkdtempSync(join(tmpdir(), "automation-captcha-campaign-"));
-  const ledgerDir = join(rootDir, "ledger");
-  const binDir = join(rootDir, "bin");
-  const launchPath = join(rootDir, "launches.log");
-  const previousCwd = process.cwd();
-  const previousPath = process.env.PATH;
-  const previousLaunchPath = process.env.CAPTCHA_LAUNCH_PATH;
-  const previousRoundScript = process.env.CAPTCHA_ROUND_SCRIPT;
-  try {
-    mkdirSync(binDir, { recursive: true });
-    const roundScript = writeTypedCaptchaRoundFixture(rootDir, {});
-    writeFileSync(join(rootDir, "settings.json"), "{}\n", "utf8");
-    process.chdir(rootDir);
-    process.env.PATH = `${binDir}:${previousPath ?? ""}`;
-    process.env.CAPTCHA_LAUNCH_PATH = launchPath;
-    process.env.CAPTCHA_ROUND_SCRIPT = roundScript;
-
-    assert.deepEqual(
-      await runAutomationTask("fubon-all-statements", ledgerDir),
-      { status: "failed" },
-    );
-
-    const launches = readFileSync(launchPath, "utf8")
-      .trim()
-      .split(/\r?\n/)
-      .filter(Boolean);
-    assert.equal(launches.length, 10);
-    assert.equal(new Set(launches).size, 10);
-    const db = openLedgerDatabase(ledgerDir, { readOnly: true });
-    const runs = db
-      .prepare("SELECT task_run_id, status, attempt, max_attempts FROM automation_task_runs")
-      .all() as Array<{
-        task_run_id: string;
-        status: string;
-        attempt: number;
-        max_attempts: number;
-      }>;
-    assert.equal(runs.length, 1);
-    assert.deepEqual(
-      {
-        status: runs[0]?.status,
-        attempt: runs[0]?.attempt,
-        maxAttempts: runs[0]?.max_attempts,
-      },
-      { status: "failed", attempt: 10, maxAttempts: 10 },
-    );
-    db.close();
-  } finally {
-    process.chdir(previousCwd);
-    if (previousPath === undefined) delete process.env.PATH;
-    else process.env.PATH = previousPath;
-    if (previousLaunchPath === undefined) delete process.env.CAPTCHA_LAUNCH_PATH;
-    else process.env.CAPTCHA_LAUNCH_PATH = previousLaunchPath;
-    if (previousRoundScript === undefined) delete process.env.CAPTCHA_ROUND_SCRIPT;
-    else process.env.CAPTCHA_ROUND_SCRIPT = previousRoundScript;
-    rmSync(rootDir, { recursive: true, force: true });
-  }
-});
-
-test("cancellation during CAPTCHA cleanup prevents the next workflow launch", async () => {
-  const rootDir = mkdtempSync(join(tmpdir(), "automation-captcha-cancel-"));
-  const ledgerDir = join(rootDir, "ledger");
-  const binDir = join(rootDir, "bin");
-  const launchPath = join(rootDir, "launches.log");
-  const cleanupStartedPath = join(rootDir, "cleanup-started");
-  const allowCleanupPath = join(rootDir, "allow-cleanup");
-  const previousCwd = process.cwd();
-  const previousPath = process.env.PATH;
-  const previousLaunchPath = process.env.CAPTCHA_LAUNCH_PATH;
-  const previousRoundScript = process.env.CAPTCHA_ROUND_SCRIPT;
-  let runPromise: Promise<{ status: string }> | null = null;
-  try {
-    mkdirSync(binDir, { recursive: true });
-    const roundScript = writeTypedCaptchaRoundFixture(rootDir, {
-      cleanupStartedPath,
-      allowCleanupPath,
-    });
-    writeFileSync(join(rootDir, "settings.json"), "{}\n", "utf8");
-    process.chdir(rootDir);
-    process.env.PATH = `${binDir}:${previousPath ?? ""}`;
-    process.env.CAPTCHA_LAUNCH_PATH = launchPath;
-    process.env.CAPTCHA_ROUND_SCRIPT = roundScript;
-
-    runPromise = runAutomationTask("fubon-all-statements", ledgerDir);
-    await waitForFile(cleanupStartedPath);
-    assert.deepEqual(await cancelAutomationTask("fubon-all-statements"), {
-      cancelled: "fubon-all-statements",
-    });
-    writeFileSync(allowCleanupPath, "ok\n", "utf8");
-    assert.deepEqual(await runPromise, { status: "failed" });
-
-    const launches = readFileSync(launchPath, "utf8")
-      .trim()
-      .split(/\r?\n/)
-      .filter(Boolean);
-    assert.equal(launches.length, 1);
-    const db = openLedgerDatabase(ledgerDir, { readOnly: true });
-    const runs = db
-      .prepare("SELECT status, attempt, max_attempts FROM automation_task_runs")
-      .all() as Array<{ status: string; attempt: number; max_attempts: number }>;
-    assert.equal(runs.length, 1);
-    assert.equal(runs[0]?.status, "failed");
-    assert.equal(runs[0]?.attempt, 1);
-    assert.equal(runs[0]?.max_attempts, 10);
-    db.close();
-  } finally {
-    if (runPromise) writeFileSync(allowCleanupPath, "ok\n", "utf8");
-    process.chdir(previousCwd);
-    if (previousPath === undefined) delete process.env.PATH;
-    else process.env.PATH = previousPath;
-    if (previousLaunchPath === undefined) delete process.env.CAPTCHA_LAUNCH_PATH;
-    else process.env.CAPTCHA_LAUNCH_PATH = previousLaunchPath;
-    if (previousRoundScript === undefined) delete process.env.CAPTCHA_ROUND_SCRIPT;
-    else process.env.CAPTCHA_ROUND_SCRIPT = previousRoundScript;
     rmSync(rootDir, { recursive: true, force: true });
   }
 });
