+++ src/lib/automation/server/runner.ts
@@ -2,7 +2,6 @@ import { spawnSync } from "node:child_process";
 import { openLedgerDatabase } from "../../../ledger/db/client.ts";
 import { resolvePatchCommand } from "./desktop-command.ts";
 import {
-  finalizeAutomationTaskRun,
   finalizePersistedActiveRuns,
   finalizePersistedRun,
   scheduleAutomationTaskRunTimeout,
@@ -43,14 +42,9 @@ export {
   resumeFailureMessage,
 } from "./task-run-execution.ts";
 import {
-  appendLog,
   automationSessionOwnerForRun,
-  errorMessage,
-  finalizeAutomationSessionForRun,
   isLiveOwnedAutomationSession,
   relinquishAutomationSessionForTask,
-  sessionFromRun,
-  type OwnedAutomationSession,
   type LiveAutomationSessionDependencies,
 } from "./automation-session-disposition.ts";
 import {
@@ -61,8 +55,6 @@ import {
 } from "./session-lifecycle.ts";
 import {
   activeTaskRuns,
-  taskRunById,
-  updateTaskRun,
   type AutomationTaskRun,
   type AutomationTaskStatus,
 } from "./store.ts";
@@ -72,16 +64,10 @@ import {
   selectStatementTypes,
 } from "../statement-selection.ts";
 import { readAutomationSettings } from "./settings.ts";
-import { routeWaitingRunVerification } from "./verification-routing.ts";
 import {
-  createCaptchaRetryCampaign,
-  isCaptchaRetryCampaignReady,
-  transitionCaptchaRetryCampaign,
-  type CaptchaRetryCampaign,
-} from "./captcha-retry-campaign.ts";
-import { MAX_CAPTCHA_RETRY_ROUNDS } from "./captcha-retry-campaign.ts";
-import type { CaptchaRoundOutcomeMessage } from "../captcha-round-outcome.ts";
-import type { AutomationTaskProcessResult } from "./task-run-finalization.ts";
+  runCaptchaRetryCampaign,
+} from "./captcha-retry-coordinator.ts";
+import type { AutomationTaskExecutionOptions } from "./task-run-execution.ts";
 
 export { closeLibrettoSession };
 
@@ -437,176 +423,6 @@ export async function shutdownAutomationSessions(
     throw new AggregateError(errors, "Failed to shut down automation sessions");
 }
 
-type CaptchaRetryExecutionResult = Awaited<
-  ReturnType<typeof runAutomationTaskExecution>
->;
-
-function processResultOf(execution: CaptchaRetryExecutionResult) {
-  return "result" in execution ? execution.result : null;
-}
-
-function executionIdOf(execution: CaptchaRetryExecutionResult) {
-  return "executionId" in execution ? execution.executionId : null;
-}
-
-function captchaOutcomeMessagesOf(execution: CaptchaRetryExecutionResult) {
-  return processResultOf(execution)?.captchaOutcomeMessages ?? [];
-}
-
-function retryableCaptchaOutcomeOf(execution: CaptchaRetryExecutionResult) {
-  return captchaOutcomeMessagesOf(execution)
-    .map((message) => message.outcome)
-    .findLast((outcome) => outcome.kind === "retryable") ?? null;
-}
-
-function recordCapturedChallenge(
-  campaign: CaptchaRetryCampaign,
-  executionId: string,
-) {
-  if (campaign.status === "ready") {
-    return transitionCaptchaRetryCampaign(campaign, {
-      kind: "challenge-captured",
-      executionId,
-    });
-  }
-  if (
-    campaign.status === "awaiting-outcome" &&
-    campaign.activeExecutionId === executionId
-  ) {
-    return campaign;
-  }
-  return campaign;
-}
-
-function recordTypedCaptchaMessages(
-  campaign: CaptchaRetryCampaign,
-  messages: readonly CaptchaRoundOutcomeMessage[],
-) {
-  let next = campaign;
-  for (const message of messages) {
-    if (message.outcome.kind === "captured") {
-      next = recordCapturedChallenge(next, message.executionId);
-      continue;
-    }
-    // No provider currently has an audited rejection probe. Keep this
-    // terminal signal fail-closed until a provider registers evidence; it
-    // must never advance the retry campaign on its own.
-    if (
-      message.outcome.kind === "retryable" &&
-      message.outcome.reason === "provider-rejected"
-    ) continue;
-    if (next.status !== "awaiting-outcome") continue;
-    // A resume process can have a new transport execution identity while it
-    // is still completing the challenge captured by the previous process.
-    // The authenticated socket has already proven the child identity; the
-    // campaign identity remains the active round identity.
-    const eventExecutionId = next.activeExecutionId;
-    next = transitionCaptchaRetryCampaign(next, {
-      kind: "round-outcome",
-      executionId: eventExecutionId,
-      outcome: message.outcome.kind === "retryable"
-        ? { kind: "retryable", reason: "solver-exhausted" }
-        : message.outcome.kind === "succeeded"
-          ? { kind: "succeeded" }
-          : message.outcome.kind === "cancelled"
-            ? { kind: "cancelled" }
-            : { kind: "failed", reason: "workflow-failed" },
-    });
-  }
-  return next;
-}
-
-function markCaptchaCampaignCancelled(campaign: CaptchaRetryCampaign) {
-  if (
-    campaign.status === "completed" ||
-    campaign.status === "failed" ||
-    campaign.status === "cancelled" ||
-    campaign.status === "exhausted"
-  ) return campaign;
-  return transitionCaptchaRetryCampaign(campaign, {
-    kind: "cancelled",
-    executionId: campaign.status === "awaiting-outcome"
-      ? campaign.activeExecutionId
-      : undefined,
-  });
-}
-
-async function finalizeCaptchaRetryCampaign(
-  taskDb: ReturnType<typeof openLedgerDatabase>,
-  taskRunId: string,
-  result: CaptchaRetryExecutionResult,
-  ledgerDir: string,
-  message: string,
-) {
-  const run = taskRunById(taskDb, taskRunId);
-  if (!run) return { status: "failed" as const };
-  const processResult = processResultOf(result);
-  const fallback: AutomationTaskProcessResult = {
-    exitCode: null,
-    signal: null,
-    error: new Error(message),
-    logTail: run.logTail,
-    resumeFailure: null,
-    statementSummary: null,
-    outputPersistenceWarnings: [],
-    externalPrerequisiteIds: [],
-    captchaOutcomeMessages: [],
-  };
-  const finalized = await finalizeAutomationTaskRun(
-    {
-      taskDb,
-      taskId: run.taskId,
-      taskKind: run.kind,
-      taskRunId,
-      logPath: run.logPath,
-      ledgerDir,
-    },
-    {
-      ...(processResult ?? fallback),
-      exitCode: null,
-      signal: null,
-      error: new Error(message),
-      logTail: processResult?.logTail ?? run.logTail,
-    },
-  );
-  return finalized;
-}
-
-async function cleanUpCaptchaRetryRound(
-  taskDb: ReturnType<typeof openLedgerDatabase>,
-  taskRunId: string,
-  reason: string,
-  activeOwner?: OwnedAutomationSession | null,
-) {
-  const run = taskRunById(taskDb, taskRunId);
-  if (!run) return { ok: false as const, message: "Missing automation task run." };
-  const cleanup = (activeOwner ?? sessionFromRun(run))
-    ? await finalizeAutomationSessionForRun(run, null, "exact", activeOwner)
-    : null;
-  if (cleanup?.cleanupFailed) {
-    return {
-      ok: false as const,
-      message: cleanup.errorMessage ?? "CAPTCHA retry session cleanup failed.",
-    };
-  }
-  const marker = `captcha-retry: restarting workflow (${reason})`;
-  try {
-    appendLog(run.logPath, marker + "\n");
-  } catch {
-    // The finalization path will retain the existing log-tail warning if the
-    // log cannot be appended; a retry does not depend on this diagnostic.
-  }
-  updateTaskRun(taskDb, taskRunId, {
-    status: "running",
-    finishedAt: null,
-    exitCode: null,
-    signal: null,
-    errorMessage: null,
-    logTail: `${run.logTail}\n${marker}\n`.slice(-4_000),
-  });
-  return { ok: true as const };
-}
-
 export async function runAutomationTask(
   taskId: string,
   ledgerDir = process.env.LEDGER_DIR ?? "data/ledger",
@@ -628,20 +444,13 @@ export async function runAutomationTask(
     // Capture them once so a settings-file edit while a CAPTCHA campaign is
     // between rounds cannot silently change the next execution's policy.
     const launchVerificationSettings = { ...readAutomationSettings() };
-    let campaign: CaptchaRetryCampaign = createCaptchaRetryCampaign();
     let taskRunId: string | null = null;
     const onRunCreated = (createdTaskRunId: string) => {
       taskRunId = createdTaskRunId;
       activeTaskRunIds.set(taskId, createdTaskRunId);
     };
-    type RoutedExecution = {
-      execution: CaptchaRetryExecutionResult;
-      routing?: Awaited<ReturnType<typeof routeWaitingRunVerification>>;
-    };
-    const runExecution = async (
-      executionOptions: Parameters<typeof runAutomationTaskExecution>[3],
-    ): Promise<RoutedExecution> => {
-      const execution = await runAutomationTaskExecution(
+    const execution = (executionOptions: AutomationTaskExecutionOptions) =>
+      runAutomationTaskExecution(
         task,
         db!,
         ledgerDir,
@@ -654,217 +463,19 @@ export async function runAutomationTask(
         },
         onRunCreated,
       );
-      if (!("taskRunId" in execution) || execution.status !== "waiting_for_human") {
-        return { execution };
-      }
-
-      let resumed: RoutedExecution | null = null;
-      let routing: Awaited<ReturnType<typeof routeWaitingRunVerification>>;
-      try {
-        routing = await routeWaitingRunVerification({
-          taskId,
-          taskRunId: execution.taskRunId!,
-          db: db!,
-          scheduleResume: async (session) => {
-            resumed = await runExecution({
-              ...executionOptions,
-              taskRunId: execution.taskRunId!,
-              resumeSession: session,
-            });
-          },
-          onChallengeCaptured: () => {
-            const executionId = executionIdOf(execution);
-            if (!executionId) return;
-            campaign = recordCapturedChallenge(campaign, executionId);
-            if (campaign.status === "awaiting-outcome") {
-              updateTaskRun(db!, execution.taskRunId!, {
-                attempt: campaign.activeRound,
-                maxAttempts: campaign.maxRounds,
-              });
-            }
-          },
-          settings: launchVerificationSettings,
-        });
-      } catch (error) {
-        await finalizeCaptchaRetryCampaign(
-          db!,
-          execution.taskRunId!,
-          execution,
-          ledgerDir,
-          errorMessage(error),
-        );
-        return { execution, routing: { kind: "failed" } };
-      }
-      if (routing.kind === "resumed" && resumed) return resumed;
-      return { execution, routing };
-    };
-
-    let executionOptions: Parameters<typeof runAutomationTaskExecution>[3] = {
-      scheduledAtUtc: options.scheduledAtUtc,
-      resumeSession: options.resumeSession,
-    };
-
-    while (true) {
-      const routed = await runExecution(executionOptions);
-      const execution = routed.execution;
-      const typedMessages = captchaOutcomeMessagesOf(execution);
-      campaign = recordTypedCaptchaMessages(campaign, typedMessages);
-      if (campaign.consumedRounds > 0 && "taskRunId" in execution) {
-        updateTaskRun(db, execution.taskRunId!, {
-          attempt: campaign.status === "awaiting-outcome"
-            ? campaign.activeRound
-            : campaign.consumedRounds,
-          maxAttempts: campaign.maxRounds,
-        });
-      }
-
-      if (routed.routing?.kind === "failed") {
-        return { status: "failed" as const };
-      }
-      if (execution.status === "cancelled") {
-        if ("taskRunId" in execution) {
-          await finalizeCaptchaRetryCampaign(
-            db,
-            execution.taskRunId!,
-            execution,
-            ledgerDir,
-            "Automation task cancelled.",
-          );
-        }
-        return { status: "failed" as const };
-      }
-      if (
-        automationTaskCancellationRequested(taskId) &&
-        "taskRunId" in execution
-      ) {
-        campaign = markCaptchaCampaignCancelled(campaign);
-        await finalizeCaptchaRetryCampaign(
-          db,
-          execution.taskRunId!,
-          execution,
-          ledgerDir,
-          "Automation task cancelled.",
-        );
-        return { status: "failed" as const };
-      }
-      if (routed.routing?.kind === "human") {
-        if (campaign.status === "awaiting-outcome") {
-          campaign = transitionCaptchaRetryCampaign(campaign, {
-            kind: "round-outcome",
-            executionId: campaign.activeExecutionId,
-            outcome: { kind: "succeeded" },
-          });
-        }
-        return { status: "waiting_for_human" as const };
-      }
-
-      const typedRetry = retryableCaptchaOutcomeOf(execution);
-      const routeRetry = routed.routing?.kind === "retryable"
-        ? "solver-exhausted" as const
-        : typedRetry?.reason === "solver-exhausted"
-          ? "solver-exhausted" as const
-          : null;
-      if (typedRetry?.reason === "provider-rejected") {
-        if ("taskRunId" in execution) {
-          await finalizeCaptchaRetryCampaign(
-            db,
-            execution.taskRunId!,
-            execution,
-            ledgerDir,
-            "Provider CAPTCHA rejection proof is not enabled for this workflow.",
-          );
-        }
-        return { status: "failed" as const };
-      }
-      if (routeRetry) {
-        if (!("taskRunId" in execution)) return { status: "failed" as const };
-        // A workflow-side reporter may have already delivered both the
-        // capture and solver-exhausted events before its process exits. In
-        // that case recordTypedCaptchaMessages has advanced the campaign to
-        // ready (or exhausted); a host-side solver route still arrives in the
-        // awaiting state and needs the transition below.
-        if (campaign.status === "awaiting-outcome") {
-          campaign = transitionCaptchaRetryCampaign(campaign, {
-            kind: "round-outcome",
-            executionId: campaign.activeExecutionId,
-            outcome: { kind: "retryable", reason: routeRetry },
-          });
-        } else if (campaign.status !== "ready" && campaign.status !== "exhausted") {
-          await finalizeCaptchaRetryCampaign(
-            db,
-            execution.taskRunId!,
-            execution,
-            ledgerDir,
-            "CAPTCHA retry requested without a captured challenge.",
-          );
-          return { status: "failed" as const };
-        }
-        if (!isCaptchaRetryCampaignReady(campaign)) {
-          if (campaign.status === "exhausted") {
-            await finalizeCaptchaRetryCampaign(
-              db,
-              execution.taskRunId!,
-              execution,
-              ledgerDir,
-              `CAPTCHA retry campaign exhausted after ${MAX_CAPTCHA_RETRY_ROUNDS} rounds.`,
-            );
-            return { status: "failed" as const };
-          }
-          return { status: "failed" as const };
-        }
-        const cleanup = await cleanUpCaptchaRetryRound(
-          db,
-          execution.taskRunId!,
-          routeRetry,
-          "owner" in execution ? execution.owner : null,
-        );
-        if (!cleanup.ok) {
-          await finalizeCaptchaRetryCampaign(
-            db,
-            execution.taskRunId!,
-            execution,
-            ledgerDir,
-            cleanup.message,
-          );
-          return { status: "failed" as const };
-        }
-        if (automationTaskCancellationRequested(taskId)) {
-          campaign = markCaptchaCampaignCancelled(campaign);
-          await finalizeCaptchaRetryCampaign(
-            db,
-            execution.taskRunId!,
-            execution,
-            ledgerDir,
-            "Automation task cancelled.",
-          );
-          return { status: "failed" as const };
-        }
-        executionOptions = {
-          scheduledAtUtc: options.scheduledAtUtc,
-          taskRunId: execution.taskRunId!,
-          attempt: campaign.nextRound,
-          maxAttempts: MAX_CAPTCHA_RETRY_ROUNDS,
-        };
-        continue;
-      }
-
-      if (
-        campaign.status === "awaiting-outcome" &&
-        "taskRunId" in execution &&
-        (execution.status === "completed" || execution.status === "partial")
-      ) {
-        campaign = transitionCaptchaRetryCampaign(campaign, {
-          kind: "round-outcome",
-          executionId: campaign.activeExecutionId,
-          outcome: { kind: "succeeded" },
-        });
-      }
-      return {
-        status: "status" in execution
-          ? execution.status === "captcha-retryable" ? "failed" : execution.status
-          : "failed",
-      };
-    }
+    return await runCaptchaRetryCampaign({
+      taskId,
+      taskDb: db,
+      ledgerDir,
+      launchVerificationSettings,
+      initialExecutionOptions: {
+        scheduledAtUtc: options.scheduledAtUtc,
+        resumeSession: options.resumeSession,
+      },
+      execute: execution,
+      isCancellationRequested: () =>
+        automationTaskCancellationRequested(taskId),
+    });
   } finally {
     activeTaskRunIds.delete(taskId);
     cancellationRequestedTaskIds.delete(taskId);
